Move extra data methods to separate interface.
